MakeMyPi example
MakeMyPi ( https://github.com/mholt/makemypi ) is a script that prepares an SD card for your Raspberry Pi with the recommended Raspbian Wheezy distro and does all the setup and configuration for you.

It supports running a one-time provision script that you specify, enabling Wifi, and more.

See the GitHub project page ( https://github.com/mholt/makemypi ) for more details and instructions.

It's easy to use: just clone down the repo, configure a few things, and run the script, following instructions as you go. Most of it is automatic and you can be running in just a few minutes.